Server/requirements.txt

flask==2.3.2
python-dotenv==1.0.0
supabase==2.3.0
twilio==8.12.0
requests==2.31.0